packages/omd/pyproject.toml
[project]
name = "hangar-omd"
version = "0.1.0"
description = "MDAO server: materializes analysis plans into OpenMDAO problems"
requires-python = ">=3.11"
dependencies = [
    "hangar-sdk[all]",
    "openmdao>=3.35,!=3.40",
    "openaerostruct>=2.12,<3",
    "click>=8.0",
    "jsonschema>=4.0",
    "pyyaml>=6.0",
]

[project.scripts]
omd-server = "hangar.omd.server:main"
omd-cli    = "hangar.omd.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hangar"]

[tool.uv.sources]
hangar-sdk = { workspace = true }
openaerostruct = { path = "../../upstream/OpenAeroStruct", editable = true }

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "parity: parity tests vs direct API",
    "eval: evaluation test cases",
    "golden_physics: tier 1 physics invariant tests",
    "golden_numerics: tier 2 numeric baseline tests",
]
